Randomly update supplier's prices every 5 seconds
The update happens in the following manner:
The supplier generates a random price for a car component based on the hardcoded reference prices.
The price update is usually an update that is of a variation percentage between 0 and 20%.
If the generated price falls below or above the thresholds (i.e. if the update exceeds
or falls under 50% of the reference price, the updated price is set to the corresponding threshold).

Let's assume that the reference price for an engine block is 500.0$.
For example, if an engine block currently costs 500.0$, the random number for the variation
percentage (which is between 0 and 1) is 0.5 and the variation sign is -1.0 (in other words,
negative), then the final result of the updated price will be:
     500.0 + (-1.0) * 0.5 * 0.2 * 500.0 = 450.0$
which represents a variation of -10% on the previous price.

In another example, we have another engine block that currently costs 300.0$, the random number
for the variation percentage is 1.0, which multiplied with 0.2 returns 0.2 representing a 20% variation
for the price. The variation sign is -1.0, so the result is:
     300.0 + (-1.0) * 1.0 * 0.2 * 300.0 = 240.0
But 240.0 is less than the lower threshold for the price, which is half of the reference price of 500.0,
i.e. 250.0. So, the new price will actually be set to 250.0$.